[Support] Move time related classes to new Time namespace.
