+++ src/Valkyrja/Application/Entry/App.php
@@ -27,7 +27,7 @@ use Valkyrja\Http\Message\Factory\RequestFactory;
 use Valkyrja\Http\Message\Request\Contract\ServerRequest;
 use Valkyrja\Http\Server\Handler\Contract\RequestHandler;
 use Valkyrja\Support\Directory;
-use Valkyrja\Support\Microtime;
+use Valkyrja\Support\Time\Microtime;
 use Valkyrja\Throwable\Handler\Contract\ThrowableHandler as ThrowableHandlerContract;
 use Valkyrja\Throwable\Handler\ThrowableHandler;
