+++ src/Valkyrja/Http/Message/Header/Value/Cookie.php
@@ -18,7 +18,7 @@ use Override;
 use Valkyrja\Http\Message\Enum\SameSite;
 use Valkyrja\Http\Message\Header\Value\Component\Component;
 use Valkyrja\Http\Message\Header\Value\Contract\Cookie as Contract;
-use Valkyrja\Support\Time;
+use Valkyrja\Support\Time\Time;
 
 use function array_filter;
 use function gmdate;
