+++ src/Valkyrja/Http/Middleware/Cache/CacheResponseMiddleware.php
@@ -25,7 +25,7 @@ use Valkyrja\Http\Middleware\Contract\TerminatedMiddleware;
 use Valkyrja\Http\Middleware\Handler\Contract\RequestReceivedHandler;
 use Valkyrja\Http\Middleware\Handler\Contract\TerminatedHandler;
 use Valkyrja\Support\Directory;
-use Valkyrja\Support\Time;
+use Valkyrja\Support\Time\Time;
 use Valkyrja\Throwable\Exception\RuntimeException;
 
 use function base64_decode;
