+++ tests/Unit/Http/Message/Response/ResponseTest.php
@@ -13,12 +13,13 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests\Unit\Http\Message\Response
 {
+
     use Valkyrja\Http\Message\Constant\HeaderName;
     use Valkyrja\Http\Message\Enum\StatusCode;
     use Valkyrja\Http\Message\Header\Value\Cookie;
     use Valkyrja\Http\Message\Response\Response;
     use Valkyrja\Http\Message\Stream\Stream;
-    use Valkyrja\Support\Time;
+    use Valkyrja\Support\Time\Time;
     use Valkyrja\Tests\Unit\TestCase;
 
     use function ob_get_clean;
@@ -189,6 +190,7 @@ namespace Valkyrja\Tests\Unit\Http\Message\Response
 
 namespace Valkyrja\Http\Message\Response
 {
+
     use Valkyrja\Tests\Unit\Http\Message\Response\ResponseTest;
 
     function header(string $header, bool $replace = true, int $response_code = 0): void
